:tractor: Move Handle* commands to their respective handlers
